+++ rollup.config.es6.js
@@ -1,6 +1,10 @@
 import config from './rollup.config';
+const [configIndex, configPlugin] = config;
 
-config.output.format = 'es';
-config.output.file = 'dist/i18nizeelement.es6.js';
+configIndex.output.format = 'es';
+configIndex.output.file = 'dist/i18nizeelement.es6.js';
 
-export default config;
+configPlugin.output.format = 'es';
+configPlugin.output.file = 'dist/i18nizeelement-jamilih-plugin.es6.js';
+
+export default [configIndex, configPlugin];