+++ CHANGES.md
@@ -1,3 +1,9 @@
+# 0.2.0
+
+- Build: Add i18nizeElement Jamilih plugin builds, ensuring buble runs
+    before other Rollup plugins
+- Refactoring: Rename plugin file
+
 # 0.1.0
 
 - Initial release